+++ examples/data-sources/datahub_glossary_nodes/data-source.tf
@@ -1,0 +1,5 @@
+data "datahub_glossary_nodes" "all" {}
+
+output "glossary_node_urns" {
+  value = data.datahub_glossary_nodes.all.urns
+}