+++ examples/data-sources/datahub_glossary_terms/data-source.tf
@@ -1,0 +1,5 @@
+data "datahub_glossary_terms" "all" {}
+
+output "glossary_term_urns" {
+  value = data.datahub_glossary_terms.all.urns
+}